mcp-client-gpt/pyproject.toml
[project]
name = "mcp-client-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gpt>=0.2",
    "mcp>=1.9.3",
    "openai==0.28.1",
    "python-dotenv>=1.1.0",
]
